[hotfix] Extract JobManagerTable interface + rename impl into DefaultJobManagerTable
This commit allows to provide different JobManagerTable implementations to the TaskExecutor.
