+++ flink-runtime/src/test/java/org/apache/flink/runtime/taskexecutor/TaskSubmissionTestEnvironment.java
@@ -151,7 +151,7 @@ class TaskSubmissionTestEnvironment implements AutoCloseable {
 
 		this.testingRpcService = testingRpcService;
 		final JobManagerConnection jobManagerConnection = createJobManagerConnection(jobId, jobMasterGateway, testingRpcService, taskManagerActions, timeout);
-		final JobManagerTable jobManagerTable = new JobManagerTable();
+		final JobManagerTable jobManagerTable = new DefaultJobManagerTable();
 		jobManagerTable.put(jobId, jobManagerConnection);
 
 		TaskExecutorLocalStateStoresManager localStateStoresManager = new TaskExecutorLocalStateStoresManager(
